winunpack: initial work on comfort ext traits
The 7-zip COM API almost always returns HRESULT and passes the actual
return value through pointer parameters. Write a few creature-comfort
extension traits that map the HRESULT to a Result and pass the return
value as a return value.

Most of the COM APIs in the windows crate ship with these creature
comforts, but the code generated by the #[interface] macro is rather
low-level in comparison.